py/taskapiclient/requirements.txt
python-dotenv==0.15.0
requests==2.25.1
opencensus==0.7.11
opencensus-ext-azure==1.0.5
